+++ conftest.py
@@ -1,3 +1,4 @@
+from distutils.version import LooseVersion
 import hashlib
 import os
 import shlex
@@ -164,66 +165,68 @@ def patch_nengo_tests():
 
     test_synapses.allclose = allclose
 
-    # cast output of run_synapse to float64. this is necessary because
-    # Synapse.filt bases its internal dtypes on the dtype of its inputs, and
-    # we don't want to downcast everything there to float32.
-    nengo_run_synapse = test_synapses.run_synapse
+    if LooseVersion(nengo.__version__) < "3.0.0":
+        # cast output of run_synapse to float64. this is necessary because
+        # Synapse.filt bases its internal dtypes on the dtype of its inputs, and
+        # we don't want to downcast everything there to float32.
+        nengo_run_synapse = test_synapses.run_synapse
 
-    def run_synapse(*args, **kwargs):
-        output = nengo_run_synapse(*args, **kwargs)
-        return tuple(x.astype(np.float64) for x in output)
+        def run_synapse(*args, **kwargs):
+            output = nengo_run_synapse(*args, **kwargs)
+            return tuple(x.astype(np.float64) for x in output)
 
-    test_synapses.run_synapse = run_synapse
+        test_synapses.run_synapse = run_synapse
 
-    # remove the correction probes from _test_pes
-    def _test_pes(NengoSimulator, nl, plt, seed,
-                  pre_neurons=False, post_neurons=False, weight_solver=False,
-                  vin=np.array([0.5, -0.5]), vout=None, n=200,
-                  function=None, transform=np.array(1.), rate=1e-3):
-        vout = np.array(vin) if vout is None else vout
+        # remove the correction probes from _test_pes
+        def _test_pes(NengoSimulator, nl, plt, seed,
+                      pre_neurons=False, post_neurons=False,
+                      weight_solver=False,
+                      vin=np.array([0.5, -0.5]), vout=None, n=200,
+                      function=None, transform=np.array(1.), rate=1e-3):
+            vout = np.array(vin) if vout is None else vout
 
-        model = nengo.Network(seed=seed)
-        with model:
-            model.config[nengo.Ensemble].neuron_type = nl()
+            model = nengo.Network(seed=seed)
+            with model:
+                model.config[nengo.Ensemble].neuron_type = nl()
 
-            u = nengo.Node(output=vin)
-            v = nengo.Node(output=vout)
-            a = nengo.Ensemble(n, dimensions=u.size_out)
-            b = nengo.Ensemble(n, dimensions=u.size_out)
-            e = nengo.Ensemble(n, dimensions=v.size_out)
+                u = nengo.Node(output=vin)
+                v = nengo.Node(output=vout)
+                a = nengo.Ensemble(n, dimensions=u.size_out)
+                b = nengo.Ensemble(n, dimensions=u.size_out)
+                e = nengo.Ensemble(n, dimensions=v.size_out)
 
-            nengo.Connection(u, a)
+                nengo.Connection(u, a)
 
-            bslice = b[:v.size_out] if v.size_out < u.size_out else b
-            pre = a.neurons if pre_neurons else a
-            post = b.neurons if post_neurons else bslice
+                bslice = b[:v.size_out] if v.size_out < u.size_out else b
+                pre = a.neurons if pre_neurons else a
+                post = b.neurons if post_neurons else bslice
 
-            conn = nengo.Connection(pre, post,
-                                    function=function, transform=transform,
-                                    learning_rule_type=nengo.PES(rate))
-            if weight_solver:
-                conn.solver = nengo.solvers.LstsqL2(weights=True)
+                conn = nengo.Connection(pre, post,
+                                        function=function, transform=transform,
+                                        learning_rule_type=nengo.PES(rate))
+                if weight_solver:
+                    conn.solver = nengo.solvers.LstsqL2(weights=True)
 
-            nengo.Connection(v, e, transform=-1)
-            nengo.Connection(bslice, e)
-            nengo.Connection(e, conn.learning_rule)
+                nengo.Connection(v, e, transform=-1)
+                nengo.Connection(bslice, e)
+                nengo.Connection(e, conn.learning_rule)
 
-            b_p = nengo.Probe(bslice, synapse=0.03)
-            e_p = nengo.Probe(e, synapse=0.03)
+                b_p = nengo.Probe(bslice, synapse=0.03)
+                e_p = nengo.Probe(e, synapse=0.03)
 
-            weights_p = nengo.Probe(conn, 'weights', sample_every=0.01)
+                weights_p = nengo.Probe(conn, 'weights', sample_every=0.01)
 
-        with NengoSimulator(model) as sim:
-            sim.run(0.5)
-        t = sim.trange()
-        weights = sim.data[weights_p]
+            with NengoSimulator(model) as sim:
+                sim.run(0.5)
+            t = sim.trange()
+            weights = sim.data[weights_p]
 
-        tend = t > 0.4
-        assert np.allclose(sim.data[b_p][tend], vout, atol=0.05)
-        assert np.allclose(sim.data[e_p][tend], 0, atol=0.05)
-        assert not np.allclose(weights[0], weights[-1], atol=1e-5)
+            tend = t > 0.4
+            assert np.allclose(sim.data[b_p][tend], vout, atol=0.05)
+            assert np.allclose(sim.data[e_p][tend], 0, atol=0.05)
+            assert not np.allclose(weights[0], weights[-1], atol=1e-5)
 
-    test_learning_rules._test_pes = _test_pes
+        test_learning_rules._test_pes = _test_pes
 
 
 patch_nengo_tests()