+++ nengo_dl/tests/test_processes.py
@@ -55,6 +55,22 @@ def test_alpha_multidim(Simulator, seed):
         assert np.allclose(sim.data[p1], canonical[1], atol=5e-5)
 
 
+def test_linearfilter(Simulator, seed):
+    # The following num, den are for a 4th order analog Butterworth filter,
+    # generated with `scipy.signal.butter(4, 0.1, analog=False)`
+    butter_num = np.array(
+        [0.0004166, 0.0016664, 0.0024996, 0.0016664, 0.0004166])
+    butter_den = np.array(
+        [1., -3.18063855, 3.86119435, -2.11215536, 0.43826514])
+
+    dt = 1e-3
+    synapse = LinearFilter(butter_num, butter_den, analog=False)
+    t, x, yhat = run_synapse(Simulator, seed, synapse, dt=dt)
+    y = synapse.filt(x, dt=dt, y0=0)
+
+    assert allclose(t, y, yhat, delay=dt, atol=1e-4)
+
+
 def test_linear_analog(Simulator, seed):
     dt = 1e-3
 
@@ -94,7 +110,7 @@ def test_zero_matrices(Simulator, zero, seed):
     t, x, yhat = run_synapse(Simulator, seed, synapse, dt=dt)
     y = synapse.filt(x, dt=dt, y0=0)
 
-    assert allclose(t, y, yhat, delay=dt * 2 if zero == "D" else dt, atol=5e-5)
+    assert allclose(t, y, yhat, delay=dt, atol=5e-5)
 
 
 @pytest.mark.training
@@ -107,3 +123,56 @@ def test_linear_filter_gradient(Simulator):
 
     with Simulator(net) as sim:
         sim.check_gradients()
+
+
+def test_linearfilter_onex(Simulator):
+    with nengo.Network() as net:
+        inp = nengo.Node(lambda t: np.sin(t*10))
+
+        tau = 0.1
+
+        # check that linearfilter and lowpass are equivalent
+        # (two versions of each to check that merging works as expected)
+        p_lowpass0 = nengo.Probe(inp, synapse=tau)
+        p_lowpass1 = nengo.Probe(inp, synapse=tau * 2)
+        p_linear0 = nengo.Probe(inp, synapse=LinearFilter([1], [tau, 1]))
+        p_linear1 = nengo.Probe(inp, synapse=LinearFilter([1], [tau * 2, 1]))
+
+    with Simulator(net) as sim:
+        sim.run(0.1)
+
+        assert np.allclose(sim.data[p_lowpass0], sim.data[p_linear0])
+        assert np.allclose(sim.data[p_lowpass1], sim.data[p_linear1])
+
+
+@pytest.mark.parametrize("synapse", (
+    LinearFilter([0.1], [1], analog=False),  # NoX
+    LinearFilter([1], [0.1, 1]),  # OneX
+    Alpha(0.1),  # NoD
+    LinearFilter(
+        [0.0004166, 0.0016664, 0.0024996, 0.0016664, 0.0004166],
+        [1., -3.18063855, 3.86119435, -2.11215536, 0.43826514]),  # General
+))
+def test_linearfilter_minibatched(Simulator, synapse):
+    run_time = 0.1
+    mini_size = 4
+
+    with nengo.Network() as net:
+        inp = nengo.Node([0])
+
+        p0 = nengo.Probe(inp, synapse=synapse)
+        p1 = nengo.Probe(inp, synapse=synapse)
+
+    with Simulator(net, minibatch_size=mini_size) as sim:
+        assert len([ops for ops in sim.tensor_graph.plan if isinstance(
+            ops[0], nengo.builder.processes.SimProcess)]) == 1
+
+        data = (np.zeros((mini_size, 100, 1))
+                + np.arange(mini_size)[:, None, None])
+        sim.run(run_time, data={inp: data})
+
+    for i in range(mini_size):
+        filt = synapse.filt(np.ones((100, 1)) * i, y0=0)
+
+        assert np.allclose(sim.data[p0][i, 1:], filt[:-1])
+        assert np.allclose(sim.data[p1][i, 1:], filt[:-1])